+++ contracts/BurnableCrowdsaleToken.sol
@@ -9,7 +9,7 @@ import "./CrowdsaleToken.sol";
  */
 contract BurnableCrowdsaleToken is BurnableToken, CrowdsaleToken {
 
-  function BurnableCrowdsaleToken(string _name, string _symbol, uint _initialSupply, uint _decimals, bool _mintable)
+  function BurnableCrowdsaleToken(string _name, string _symbol, uint _initialSupply, uint8 _decimals, bool _mintable)
     CrowdsaleToken(_name, _symbol, _initialSupply, _decimals, _mintable) {
 
   }
