+++ contracts/FractionalERC20.sol
@@ -7,6 +7,6 @@ import './ERC20.sol';
  */
 contract FractionalERC20 is ERC20 {
 
-  uint public decimals;
+  uint8 public decimals;
 
 }
